Partial fix for #1087. Adding a new module hook is now working.
