+++ templates/backOffice/default/module-hooks.html
@@ -209,7 +209,7 @@ form_error_message = $form_error_message
 
     dialog_id       = "delete_module_dialog"
     dialog_title    = {intl l="Remove a module from a hook"}
-    dialog_message  = {intl l="Do you really want to remove this module ?"}
+    dialog_message  = {intl l="Do you really want to remove this module from this hook ?"}
 
     form_action         = {token_url path='/admin/module-hooks/delete'}
     form_content        = {$smarty.capture.delete_module_dialog nofilter}
